==> Observational Studies/Representation Among US Elected Officials/Project Results/Datasets and Notebooks/1_Elected_Officials_Import.ipynb <==
## Importing and Cleaning the Elected Officials Data Set

This notebook imports and cleans the elected officials data set.

# Import statements

import pandas as pd

# Reads in the ELOF dataset
df_elof = pd.read_csv('RDC-Americas-Cities-2020-Raw-Data - Raw Data- Top 100 Cities.csv')

# Converts all variables to lowercase
df_elof.columns = df_elof.columns.str.replace('\s+', '_').str.lower()

df_elof.head()

# Now we move on to restricting the data set to keep
# the relevant columns for our research questions

# Keeps only columns of relevant variables
df_elof = df_elof[["state", "race", "sex"]]

# Racial categories which will be dropped from the dataset
# due to not having an equivalent in the census data
dropped_categories = ["Other", "Unknown"]

# Drops rows which have race value in a dropped category
df_elof = df_elof[~df_elof["race"].isin(dropped_categories)]

# Placeholder, assigns a count value of one to every official
df_elof["elof_count"] = 1

# Creates the groups that the data will be binned into
columns_to_groupby = ["state", "race", "sex"]
df_elof = df_elof.groupby(by = columns_to_groupby)

# Unpacks the pandas groupby object and sums across columns which
# are not being grouped
df_elof = df_elof.sum().reset_index()

# Creates the groups that the data will be binned into
# for calculating elof_count for sex = Total
columns_to_groupby = ["state", "race"]

# Temporary df to hold the counts of elected officials
# when sex = Total
df_sex_total = df_elof.groupby(by = columns_to_groupby)

# Unpacks the pandas groupby object and sums across columns which
# are not being grouped
df_sex_total = df_sex_total.sum().reset_index()

# Adds sex column so it matches the original dataframe
df_sex_total["sex"] = "Total"

# Adds the total sex df back to the full dataframe
df_elof = df_elof.merge(df_sex_total, how = "outer")

# Sorts dataframe by state again
df_elof = df_elof.sort_values("state")

# creates the groups that the data will be binned into
# for calculating elof_count for race = Total
columns_to_groupby = ["state", "sex"]
# temporary df to hold the counts of elected officials
# when race = Total
df_race_total = df_elof.groupby(by = columns_to_groupby)
# unpacks the pandas groupby object and sums across columns which
# are not being grouped
df_race_total = df_race_total.sum().reset_index()
# adds race column so it matches the original dataframe
df_race_total["race"] = "Total"
# add the total race df back to the full dataframe
df_elof = df_elof.merge(df_race_total, how = "outer")
# sorts dataframe by state again
df_elof = df_elof.sort_values("state")

# Gets iterable of states which are in the dataset
states = df_elof["state"].unique()

for state in states:
    # Gets the total state population by summing across
    # all race categories when the state is fixed and the
    # sex category is set to total
    state_total = sum(df_elof["elof_count"].
                   loc[(df_elof["state"] == state)
                      & (df_elof["sex"] == "Total")
                      & (df_elof["race"] == "Total")])
    # Creates new column with the percentage of the state population
    df_elof.loc[df_elof["state"] == state, "percentstate_elof"] = df_elof["elof_count"].div(state_total)
    
# Outputs new file to csv
df_elof.to_csv("elof_data_cleaned.csv")

df_elof[df_elof.state == "AL"]
