==> Observational Studies/Representation Among US Elected Officials/Project Results/Datasets and Notebooks/3_Data_Merging.ipynb <==
## Merging the Census Data Set with the Elected Officials Data Set

This notebook merges the Census data set with the Elected Officials Data Set

# Import Statements

import pandas as pd

# Import the two data sets
df_pop = pd.read_csv("census_data_cleaned.csv")
df_elof = pd.read_csv("elof_data_cleaned.csv")

# Columns which are the same between the two dataframes
common_columns = ["state", "sex", "race"]

# Perform merge
df = df_pop.merge(df_elof,  on = common_columns, how = "outer")

# Multiply percentage columns' values by 100
df["percentstate_pop"] = df["percentstate_pop"].multiply(100)
df["percentstate_elof"] = df["percentstate_elof"].multiply(100)

df

# This section cleans the merged dataframe

bad_columns = ["Unnamed: 0_x", "Unnamed: 0_y"]

# Drops column with old indices now not being used
df = df.drop(columns = bad_columns)

# Replaces all NaN values with 0
df = df.fillna(0)

# Outputs new file to csv
df.to_csv("merged_data.csv")

df
