==> Observational Studies/Representation Among US Elected Officials/Project Results/Datasets and Notebooks/5_Elected_Officials_Data_for_Circle_Map.ipynb <==
## Formatting the Elected Officials Data Set for Circle Map Visual

This notebook starts with the original elected officials raw data and formats it to prepare to merge it with another data set that contains latitude and longitude coordinates for cities in the U.S. (This notebook should be run **before** the "Merge_for_Circle_Map" Jupyter Notebook.)

**NOTE: This file requires pandas version 1.1.0 to run without errors.**

# Import Statements 

import pandas as pd
pd.__version__

# Reads in the ELOF dataset
df_elof = pd.read_csv('RDC-Americas-Cities-2020-Raw-Data - Raw Data- Top 100 Cities.csv')

# Converts all variables to lowercase
df_elof.columns = df_elof.columns.str.replace('\s+', '_').str.lower()

df_elof.head()

def city_counts_df(by_state, state):
    
    '''This function takes a by_state groupby
       object and a state string and outputs a dataframe
       that includes a column with the count of 
       the number of data points for the state.
       
       by_state is includes the columns, 'jurisdiction'
       and 0, where 0 is the count of data points per jurisdiction
       
    '''
    new_df = pd.DataFrame(by_state.get_group(state).value_counts())
    
    # For some reason, this makes the index, 'state'
    new_df = new_df.reset_index('jurisdiction')
    
    # Renaming index and columns
    new_df = new_df.rename(columns={0:'count', 'jurisdiction':'city'})
    new_df = new_df.rename_axis('state')
    
    # Returns new_df
    return new_df

# Grabs all unique states in ELOF data set
states_list = list(df_elof.state.unique())

# Creates a Groupby object from ELOF DataFrame
by_state = df_elof.groupby(['state'])[['state', 'jurisdiction']]

# Initializes a DataFrame
all_cities = pd.DataFrame()

# Appends all cities and jurisdiction value_counts
# for each state to a new DataFrame
for state in states_list:
    new_df = city_counts_df(by_state, state)
    # Appends current DataFrame to existing 
    all_cities = all_cities.append(new_df)

all_cities.head()

# Now we would like to remove the word 'city'
# or 'county' from each item in the column, 'city'

def remove_text(text):
    
    '''This function takes in a state
       string from the above by_state groupby
       object and outputs the unique jurisdictions
       
    '''

    # Removing 'city' and 'county' from jurisdiction names
    city = text.lower().find('city')
    county = text.lower().find('county')
    if city != -1:
        text = text[:city - 1]
    elif county != -1: 
        text = text[:county - 1]
    else:
        pass
    
    return text

# Apply the function to all items in the 'Jurisdiction' column
all_cities['city'] = all_cities['city'].apply(lambda x: remove_text(x))
all_cities

# Looking at summary for the column, count
all_cities['count'].describe()

# Save to a csv
all_cities.to_csv('elof_cities.csv')
