==> Observational Studies/Representation Among US Elected Officials/Project Results/Datasets and Notebooks/6_Merge_for_Circle_Map.ipynb <==
## Merging Cities Data Set with Elected Officials Data Set for Initial Visual

This notebook merges **uscities.csv** and **elof_cities.csv** to plot the data from the elected officials data set onto the U.S. map. The data set, **uscities.csv**, contains latitude/longitude coordinates of cities in the U.S., which are needed for this particular graphing method. This data set was found at **https://simplemaps.com/data/us-cities**.



# Import Statements

!pip install plotly

import pandas as pd
import numpy as np
from scipy import stats
import plotly.graph_objects as go
import plotly.express as px
import matplotlib.pyplot as plt

# Reads in the cities dataset
df_cities = pd.read_csv('uscities.csv')

# Converts all variables to lowercase
df_cities.columns = df_cities.columns.str.replace('\s+', '_').str.lower()

# Displays header
df_cities.head()

# Keeps only the necessary columns
df_cities_only = df_cities[['city', 'lat', 'lng', 'population', 'state_id']]

# Renames the state column for merging purposes
df_cities_only = df_cities_only.rename(columns = {'state_id': 'state'})
df_cities_only

# Read in the elof cities dataset
df_elof_cities = pd.read_csv('elof_cities.csv')

# Converts all variables to lowercase
df_elof_cities.columns = df_elof_cities.columns.str.replace('\s+', '_').str.lower()

# Displays header
df_elof_cities.head()

# Checks for any duplicate cities
df_elof_cities[df_elof_cities.city.duplicated()]

# Since San Diego and New York are the only 
# duplicated cities, we will removes these
# rows and add the count to their original row
df_elof_cities = df_elof_cities.drop_duplicates(subset = ['city'])
df_elof_cities.city

# Note that we are losing a count of 2 for each state when
# we do this. This would be a fix we would implement in the future.

# Merge the two DataFrames
merged_cities = df_elof_cities.merge(df_cities_only, how='inner')

# Save new merged DataFrame as a csv
merged_cities.to_csv('merged_cities.csv')

merged_cities.sort_values('count')

# Reviewing the new city list
# Only has 93 items - another fix for the future
# Text most likeley did not line up exactly for each column 

city_list = []
for item in merged_cities.city:
    city_list.append(item)
    
len(city_list)

## Elected Officials Data Set Map Visualization

The outline of this code was originally taken from an example on the plotly website: https://plotly.com/python/bubble-maps/. The code was modified slightly to match our dataframe / data set.

# Used the example from https://plotly.com/python/bubble-maps/ to format this

merged_cities['text'] = (merged_cities['city']
                         + '<br>Elected Officials Count: ' + merged_cities['count'].astype(str)
                         + '<br>Population: ' + (round(merged_cities['population']/1e6, 2)).astype(str)+' million')

# 'Buckets' used to separate elected officials counts by city
# These were based on percentiles that the data fell into
limits = [(5,7),(8,9),(10,13),(14,56)]

# Associated colors with the buckets from above
colors = ['royalblue','crimson','lightseagreen','orange']

fig = go.Figure()

for i in range(len(limits)):
    lim = limits[i]
    merged_cities_sub = merged_cities[(merged_cities['count'] >= lim[0])
                                      & (merged_cities['count'] <= lim[1])]

    # Plotting on top of the current fig
    # for each loop
    fig.add_trace(go.Scattergeo(
        locationmode = 'USA-states',
        lon = merged_cities_sub['lng'],
        lat = merged_cities_sub['lat'],
        text = merged_cities_sub['text'],
        marker = dict(size = merged_cities_sub['count']*10,
                      color = colors[i],
                      line_color='rgb(40,40,40)',
                      line_width=0.5,
                      sizemode = 'area'),
        name = '{0} - {1}'.format(lim[0],lim[1])))


fig.update_layout(
        # title_text = '2019 Elected Officials Data Count',
        showlegend = True,
        margin={"r":0,"t":0,"l":0,"b":0}, #adjust window size
        geo = dict(scope = 'usa',
                   landcolor = 'rgb(217, 217, 217)')
    )

fig.show()
